pyproject.toml
[build-system]
requires = ["poetry>=1.0"]
build-backend = "poetry.masonry.api"

[tool.poetry]
name = "tron_asset_transfer"
version = "0.1.0"
description = "Helpers for making tron transactions"
license = "MIT"
authors = ["Vladimir Kamarzin <vvk@vvk.pp.ru>"]
readme = "README.md"
repository = "https://github.com/bitphage/tron-asset-transfer"

keywords = []

classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Operating System :: OS Independent",
  "Topic :: Software Development :: Libraries :: Python Modules",
]

[tool.poetry.dependencies]
# See https://python-poetry.org/docs/dependency-specification/
python = "^3.8"
tronpy = "^0.2.6"
click = "^8.1.3"
pydantic = "^1.10.4"
pyyaml = "^6.0"

[tool.poetry.dev-dependencies]
pre-commit = "^2.2.0"
pytest = "^6"
pytest-cov = "^2.7"
pytest-mock = "^3.1.0"

[tool.black]
line-length = 120
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.tox
  | \venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 120

[tool.pytest.ini_options]
# Directories that are not visited by pytest collector:
norecursedirs = "*.egg .eggs dist build docs .tox .git __pycache__"

# Extra options:
addopts = [
  "--strict-markers",
  "--doctest-modules",
  "--cov=tron_asset_transfer",
  "--cov-report=term:skip-covered",
  "--cov-report=html",
  "--cov-branch",
  "--cov-append",
  ]
